+++ src/admin/custom-block-style/item/body-area/property-label.js
